Move SLESolverTypeParser
